Switch to int scores

+++ src/eval/table_eval.h
@@ -1,9 +1,10 @@
 #ifndef TABLE_EVAL_H
 #define TABLE_EVAL_H
 
+#include <stdint.h>
 #include "../board/board.h"
 
-float table_eval(board_t *b, int c);
+int16_t table_eval(board_t *b, int c);
 int endgame_eval(board_t *b, int c);
 
 #endif
