+++ src/search/search.h
@@ -1,12 +1,13 @@
 #ifndef SEARCH_H
 #define SEARCH_H
 
+#include <stdint.h>
 #include "../board/board.h"
 
 /* Struct so that searches can return both a move and its score. */
 typedef struct {
     int pos; /* squared piece was placed */
-    float score; /* minimax score */
+    int16_t score; /* minimax score */
     int end; /* if leads to endgame */
 } move_score_t;
